rename footer button component, clean up
